requirements.txt
streamlit
googletrans==4.0.0-rc1
gtts-2.5.4
